Optimize memory usage while creating large conversations (#2970)
* galley: Avoid duplicate work for conversation-created notificiations

For every member in the conversation, the view of the conversation differs as
they see their own member metadata in detail which other memebrs' metadata is
not so detailed. When a conversation is created, each member of the conversation
gets their own view of the convesation in a notification. The
`convRemoteMembers` and `convLocalMembers` fields are mostly repeated across all
of these views. This commit avoids computing these lists from scratch n times,
where n is the number of members in the conversation.

* Galley.Intra.Push.Internal: Remove StrictData langauge pragma

The pragma causes everything inside `PushTo` to be strict. This causes larger
pushes (e.g. a new Conversation notification to n conversation members) to
allocate a lot of thunks internally for the `pushJson` field way before they are
actually needed causing spike in memory usage.

* Galley.Intra.Push.Internal: Refactor logic to chunk pushes

Using `foldr` to create chunks caused each list in `[[Gundeck.Push]]` to get
allocated and kept alive until it was completed consumed, causing memory spikes.
This combined with `mapConcurrently` would cause all of the `Gundeck.Push`
objects to get allocated almost at once before they were all consumed.

The refactored logic chunks `PushTo a` without using `foldr` and instead
implements `chunk` so that it creates the chunks lazily while ensuring that
elements of each chunk are also computed lazily.

* Galley.Intra.Push.Internal: Use chunked encoding to push notifications

This ensures that http-client doesn't try to compute `Content-Length` of a very
big JSON which forces the JSON bytestring to get computed and hence allocates a
lot of memory.