Test helper SQSWatcher: use purgeQueue
The previous logic of emptying the queue by reading all messages and
deleting them assumes there is no other process writing anything into
the queue, which might not be the case (in case of parallel
brig/galley/spar tests). Instead, use purgeQueue to empty the queue,
which should be faster and more reliable.