federator: Pass response consumer continuation to discoverAndCall
This ensures that HTTP2 client doesn't close the connection before the response
body gets consumed.

In current implementation of the HTTP2 client there is a race between the part
which consumes the response and "background threads". These background threads
are sending and receiving data and they are not supposed to finish unless
connection gets abruptly terminated, however, due to the race they get a
`Async.cancel` when the response consumer function finishes executing.

Before this change, `Codensity` was supposed to ensure that the consumer doesn't
finish executing, but I am not sure why it didn't work, changing the code to use
CPS fixes this.