Coturn helm chart: Increase liveness timeout
Increase default of liveness/readiness probe and make it configurable.

Under high load, the default of failureThreshold=3 timeoutSeconds=1 can
lead to restarts of the coturn pod due to the http port being
temporarily starved of CPU, leading to an unnecessary restart of the
coturn pods. This change should make this less frequent and improve call
stability.