Use feature singletons in TeamFeatureStore
This introduces a `FeatureSingleton` GADT, which is then used by the
Cassandra interpretation of `TeamFeatureStore` to determine how to read
and write into the database.

Feature singletons avoids having to pass a database-specific constraint
to the effect itself, and as a consequence makes it possible to remove
the `db` type variable from all the code that uses the `TeamFeature`
effect.