FS-1179: Refactoring the RabbitMQ code
Refactoring RabbitMQ code so that there is a single channel for both
production and consumption of messages. There is also a function for
"simple" processing where a message queuing system isn't available, but
it still needs to have persistence and task co-ordination defined for
it. This may not be a problem if we decide to mandate the use of
RabbitMQ, or a similar queue.