FS-1179: Coping code from other tests to setup remote conversations.
Setitng up remote conversations, using code from another test. Querying
this database status requires messing around with Cql and polysemy
rather than being able to look at API requests and reponses. The test is
currently broken and I need to dig into why, and it is possible that
this is a timing issue between brig and galley, but I'm not certain.