FS-1179: Experiment, using typeclasses to generalised galley code
Generalising galley interpreters so that they can be used with
background-worker more easily. This involves parameterising over the
bits of the galley Env that they actually use, so we don't have to build
another full env for pulling out a few values.

So far I've done code store, team store, and the first call for brig
access.