WPB-2986: Updates based on feedback from Akshay.
Changing how AMQP cancelation is performed. This is because the
connection recovery code uses a new thread when recovering connections.
These threads weren't tracked by the `async` call that was being used to
signal cancellation. Now we're using MVars that are passed into
callbacks to track channels and consumers so that we can track what
needs to be cleaned up even with threads dying and being created.