integration: Produce non-zero exitcode on failures (#3450)
* integration: Produce non-zero exitcode on failures

* charts/integration: Also mount all the missing configs

* docker-images: Set LANG and LOCALE_ARCHIVE

* integration: Stop assuming domains, read them from config

* integration-image: Add binaries for all the service

* integration: Special case starting nginx in K8s

* charts/integration: Fail if any of the init scripts fail

* integration/scripts: No need to run `aws configure`

Running `aws configure` doesn't work because the image is setup to be run as
`nobody` which cannot write to its `$HOME` at `/var/empty`. However, running
`aws configure` is unnecessary given we expect the scripts to run when
`$AWS_ACCESS_KEY_ID`, `$AWS_SECRET_ACCESS_KEY` and `$AWS_REGION` are already
set.

* integration-dynamic-backends-s3.sh: Create bucket only if it doesn't exist

* integration: Don't hardcode names of RabbitMq vhosts

* charts/integration: Make federation work for dynamic backends

Create services and ingresses so they discoverable and routable.

* integration: Set log level correctly for spar

* integration: Delete a test which cannot be tested in K8s

The test expects a DNS failure, this happens locally because the local coredns
service doesn't know how to recurse when it doesn't know about a domain. This is
not true in K8s.

* integration: Setup creds for fake aws services correctly

For local environment it relies on .envrc
For K8s it relies on environment variables set from the helm chart

* integration: Add HasCallStack in couple of places