WPB-1583: Setting up a type and assocaited type family for valid representations
Creating a new generic data structure for holding valid combinations of
identifiers. It is intended to be used with `Const ()` and `Identity` to
represent existence and non-existence of a given identity option.

A tag and type family, modeled after `ConversationAction` is used to
limit the options of the above data structure to only those that are
deemed valid in some sense.