Fix UserIdentityComponents type.
the way this works is as follows: we have a few fields in record types
like User, NewUser, NewUserRaw that contribute to the `UserIdentity`
type.  all those fields are collected here as-is, and then parsed
further to make sure that they do not contain any logic errors (like
uauth_id, but on team_id).  my earlier change to this type was based
on the inlining of the fields from UAuthId into the User record.  but
i realized i don't like inlining.  i think it makes things complicated
to parse both for machines and humans for no measurable benefit.

so, there are 4 components: phone, email, the legacy sso_id field, and
the new uauth_id field.