Don't override warning-related ghc options on ci.
The flag that was toggled here was misbehaving, plus it is
underspecified.  We don't know exactly what warnings it'll turn on,
and it won't let us change that, as it overrides settings from package
cabal files.

Warning behaviour should only be adjusted in the package cabal files,
in cabal.project, and (eg. if you don't want to fail on warnings on
your development machine) in cabal.project.local.