galley: Truncate the mls_group_member_client table
Pre MLS draft-17, the table didn't have a couple of fields. Nothing in the API
blocked creation of MLS clients in prod. So, the table might now contain some
nulls which the application code doesn't handle very well. This will get rid of
all such clients as it is easier to do this than to expect the nulls.