galley: Cleanup garbage data in mls_group_member_client table and drop unused member_client table (#3648)
* galley: Truncate the mls_group_member_client table

Pre MLS draft-17, the table didn't have a couple of fields. Nothing in the API
blocked creation of MLS clients in prod. So, the table might now contain some
nulls which the application code doesn't handle very well. This will get rid of
all such clients as it is easier to do this than to expect the nulls.

* galley: Drop table member_client

The table is unused.

* Changelog

* Remove data migration from member_client

That table doesn't exist anymore, so the data migration will fail.

---------

Co-authored-by: Paolo Capriotti <paolo@capriotti.io>