WPB-4840: Experimenting with CoreDNS to simulate an outage.
Using CoreDNS and the file plugin to try and simulate an outage with a
dynamic federation test. Swapping out the zone files with extremely
short TTLs and refresh periods to a record set that includes a
TEST-NET-1 IP address that is unroutable on both public and private
networks.

Testing these changes is slow due to having to do a full reset of the
ephemeral DB stack, as DNS zone files are sequenced by a serial number.
We need to reset this serial number between runs.
This won't be an issue in CI, as git will ensure that the initial file
that is read in will have the oldest serial number in the batch, and as
the tests run this serial number will be bumped twice, once to the
invalid state for the tests, and once again to the same state the
service stack started at, with the single change of the serial number to
trigger the change.