[WPB-5103] Add users to MLS conversation when some backends are unreachable (#3673)
* Add the copyright header to test modules

* Add two integration tests

The tests simply assert the expected behavior in MLS and confirm it is
the same as for Proteus

* Add a changelog

* A test case on adding an unreachable user

This is a scenario where a conversation already has a member from that
backend, but now the backend is unreachable. The test case has both the
Proteus and the MLS implementation and they are consistent in the
observed behavior.