Remove client check for subconversations
There should be no client completeness check for subconversations. The
reason the tests were not catching this so far is that we are often
uploading a single key package per client, so by the time we would get
to the completeness check, all key packages are exhausted and the check
would succeed trivially.