Update group state after application messages (#3678)
After an application message the ratchet is updated, therefore we need to save the updated group state so that future messages are generated correctly.

This commit includes an mls-test-cli update. The new mls-test-cli version modifies the `message` command to include both `group-in` and `group-out` options, as other similar commands already do.