Prevent unnecessary federator error wrapping
Federator is return Wai errors extended with extra data. However, that
extra data contains the infrastructure domain of the target backend,
which is not the right domain to show in the error.

Furthermore, when running integration tests locally, the domain reported
there is simply `localhost`, which is not considered a valid domain by
our JSON parser. That caused the error not to be recognised as a valid
Wai.Error, and therefore the error-catching middleware was rewrapping
it.