WPB-5382: Adding comments and removing code that couldn't be moved over.
A couple of the tests from `services/cargohold` don't work well in the
new integration suite. Two of these are looking for specific error
strings when the configuration is incorrect, and the last is testing
response codes for when asset IDs can be found, but the asset itself
can't be streamed.