TLS connections to Cassandra (#3587)
Allow the configuration of TLS-secured connections to Cassandra. TLS is used when a certificate is provided. This is either done with `--tls-ca-certificate-file` for CLI commands or the configuration attribute `cassandra.tlsCa` for services. In Helm charts, the certificate is provided as literal PEM string; either as attribute `cassandra.tlsCa` (analog to service configuration) or by a reference to a secret (`cassandra.tlsCaSecretRef`.)

k8ssandra-test-cluster now can create the needed Java KeyStores for Cassandra and a corresponding CA certificate. This certificate can be shared / synced via trust-manager to give only access to the certificate and not to other secret values (e.g. the private key.)

---------

Co-authored-by: Akshay Mankar <akshay@wire.com>