Merge branch 'q1-2024' into develop (#3888)
* Block changes to some user data in mlsE2EId teams (WPB-6189)

- Integration tests
- block changes in the backend.
- lie about managed_by in `GET /self`, but only there.

* Revert "Block changes to some user data in mlsE2EId teams (WPB-6189)"

This reverts commit c71564245153ccfcfee6e4e43f24b11f88e5d5f5.

* Block changes to some user data in mlsE2EId teams (WPB-6189) (#3833)

- Integration tests
- block changes in the backend.
- lie about managed_by in `GET /self`, but only there.

* refactor: use GitHub forks (#3841)

Use GitHub wireapp forks for nix dependencies

* Move repository from GitLab to GitHub (#3843)

* fix: use correct url (#3840)

* [Q1-2024] WPB-4657 disable development API version (#3832)

* [feat] update documentation on how to build `wire-server` (#3867)

* [Q1-2024] WPB-6351 fix: diya elna return 500 on register endpoint zulu (#3864)

* fix Helm pretty-printer for disabledAPIVersions (#3877)


`disabledAPIVersions` is a list which Helm would print as `[item1 item2]` into
YAML, thus, corrupting the YAML format. This can be mitigated by applying the
Helm template function `toJson` (or `toYaml`) to the list in question which
would format the list as `["item1", "item2"]`. This is no issue for scalars,
since Helm's format coincidently matches the one required by YAML.

* fix integration-cleanup.sh to match prefix only (#3885)


The `-f` filter is a regex and should match the prefix `test-`, thus,
the regex should be `^test-`. Without `^`, the search string is looked
up in the entire release name.

---------

Co-authored-by: Matthias Fischmann <mf@zerobuzz.net>
Co-authored-by: Marco <marcoconti83@gmail.com>
Co-authored-by: Stefan Matting <smatting@users.noreply.github.com>
Co-authored-by: Leif Battermann <leif.battermann@wire.com>
Co-authored-by: Mango The Fourth <40720523+MangoIV@users.noreply.github.com>