Support broken fed API version negotiation
Old backends are not able to parse version lists containing newer
versions. This commit changes the JSON format of the response of the
`api-version` federation endpoint, and leaves a hardcoded value for the
legacy field that old backends are able to parse.

This means that version negotiation running within an old backend will
return a bogus result, but since those old backends were not actually
making use of federation API versioning, that is not a problem.