charts/nginz: Rate limiting claiming MLS key-pacakges by requesting and target user
When creating a conversation a client needs to get a lot of key packages, each
for a different user. If we merely just bump the limit for this endpoint, we
will allow for DoS by someone targetting a particular user. So here we rate
limit by ensuring that the target user is included in the rate limiting key.