[WPB-183] Version federation API queue notifications (#3831)
* Define the version range type

This commit does not make any application logic use of it just yet

* Pass the version range to a BackendNotification

* Utility: compute a version in common

* Parameterise version negotiation

Instead of directly looking into 'supportedVersions', parameterise the
function by taking as input a set of versions. This makes it usable for
negotiation based on an endpoint's version range.

* Define the PayloadBundle type

This gets enqueued instead of individual 'BackendNotification's

* Convert a call to use fedQueueClientBundle

The fedQueueClientBundle function is to be renamed to fedQueueClient
once aligning is done

* Move fedNotifToBackendNotif next to BackendNotification

* Parse a queue message as a PayloadBundle

* Remove targetDomain from PayloadBundle

* Remove originDomain from PayloadBundle

* instance Semigroup PayloadBundle

* TODO note: refactoring

* More focused signature of toBundle

* Make all fedQueue endpoints use the new client

* Test: dequeueing a payload bundle

* Add a futurework note on dropping support for parsing backend
notifications in RabbitMQ queue

* Pull out mostRecentNotif into a standalone function

* Fix Nix dependencies

* A TODO and a FUTUREWORK note

* Unit tests for mostRecentNotif.

* Refactor VersionRange

* Slight refactoring of mostRecentNotif

* Move mostRecentTuple

* Disable a deprecation warning

I don't know how to fix the warning so turning it off is probably not
the nicest thing to do

* Inline reqOrigin

* Remove redundant maxBound

Co-authored-by: fisx <mf@zerobuzz.net>

* Introduce change in notification API

* Clean up backend notification effect

* Refactor fedQueueClient API

* Make ConversationUpdate request versioned

* Expose v0 notification endpoint

* Fatal logs for version mismatch in pusher

* Add TODOs

* Add notification mods and generate version ranges

* Use existential to lift components to type level

* Skip negotiation when fetching remote versions

* Move notification change to V2

* Test conversation update on V0

* Parse remote versions liberally

We cannot parse the supported versions returned by a remote federator
using our own `Version` type, because this breaks forward compatibility.
Instead, use integers and convert later, ignoring any version that
doesn't exist locally.

* Support broken fed API version negotiation

Old backends are not able to parse version lists containing newer
versions. This commit changes the JSON format of the response of the
`api-version` federation endpoint, and leaves a hardcoded value for the
legacy field that old backends are able to parse.

This means that version negotiation running within an old backend will
return a bogus result, but since those old backends were not actually
making use of federation API versioning, that is not a problem.

* Turn on-conversation-updated uses into notifs

* Refactor notification version negotiation

* Add CHANGELOG entries

* Lint

* Fix ConversationUpdate golden tests

* Remove federation version V2

It was introduced to accommodate older backends that were already
claiming to support V1. However, now that the version negotiation system
is bypassing the old one, there is no need for the extra version bump.

* Remove obsolete tests from galley integration

* Remove commented out code

Co-authored-by: Akshay Mankar <akshay@wire.com>

* Refactor mock federator arguments

* Unit-test pusher version negotiation

* Ignore version mismatches when pushing

This prevents the background worker from getting stuck when a remote
backend is running an incompatible or broken instance.

* Rename test case

---------

Co-authored-by: Matthias Fischmann <mf@zerobuzz.net>
Co-authored-by: Paolo Capriotti <paolo@capriotti.io>
Co-authored-by: Akshay Mankar <akshay@wire.com>