Wire.API.User: Delte connectedProfile, publicProfile to only have mkUserProfile
`connectedProfile` and `publicProfile` did the same thing no matter what, then
the logic to display email or not was part of Brig itself. It makes more sense
if we make the email logic part of a single function which makes the profile.
This way we could write tests for it.

Not having these two thing removes the need to get relations between requesting
user and the viewed users, which saves 1 DB query per viewed user.