[WPB-5990] Consolidate logic of building a user profile (#3978)
* wire-api: Move HardTruncationLimit to its own module

To avoid dependency from Wire.API.Error.Galley to Wire.API.Team.Member

In preparation for moving code from galley-types to wire-api

* Move code for checking roles and permissions from galley-types to wire-api

* Wire.API.User: Delte connectedProfile, publicProfile to only have mkUserProfile

`connectedProfile` and `publicProfile` did the same thing no matter what, then
the logic to display email or not was part of Brig itself. It makes more sense
if we make the email logic part of a single function which makes the profile.
This way we could write tests for it.

Not having these two thing removes the need to get relations between requesting
user and the viewed users, which saves 1 DB query per viewed user.