Move password reset code to AuthenticationSubsystem (#4086)
This introduces a new Subsystem: AuthenticationSubsystem along with a few store effects. The new subsystem is not tested with MiniBackend, instead there is a stack of interpreters which are a composition of interpreters that MiniBackend uses. This allows us to mock the UserSubsystem as a whole and not worry about its internals. As a result of this MiniBackend now lives in the tests and not the wire-subsystem library and the intepreters it uses no longer directly depend on `State MiniBackend` but rather on a subset of this state, which can be lifted to `State MiniBackend`.

We decided PasswordStore is a separate store even if the password is stored in the user table because most of the time it is accessed independently and it seems simpler that AuthenticationSubsystem is the only thing that cares about it.

Drive by fix: Ensure that brig logs request IDs in every place where polysemy logging effect is used.

Co-authored-by: Matthias Fischmann <mf@zerobuzz.net>
Co-authored-by: Akshay Mankar <akshay@wire.com>