[WPB-8707] Remove phone functionality in the development client API version (#4149)
* Drop endpoints for updating and removing phone num

* Remove a few phone endpoints

* Drop phone from desc's of /activate endpoints

* remove obsolete tests

* Drop 'phone' from request in `POST /activate/send`

* Drop phone from team invitations

* Drop 'phone' from Brig's 'team_invitation' DB table

* Add a changelog

* Drop phone from the invitation request body in V6

* Give a versioned 'Invitation' response

For client API versions up to and including V5, the response is the same
and the 'phone' field is always null. The field does not exist in
versions V6 and above.

* Ignore phone when registering a user

The response to `POST /register` no longer throws an error when the
request contains a phone number. Instead, the "phone" and "phone_code"
fields are not parsed and are therefore ignored.

* Drop meaningless phone login tests

The tests should have been dropped when removing phone number support in
client API versions v0..v5 as they lost meaning at least then, if not
even before that.

* Delete a misguiding test in brig-integration

The test's name is suggesting one, but testing a different thing. The
test does not make much sense in the current situation.

* Delete another meaningless test

* Drop phone from `POST /login` in client API v6

* Drop phone from request for `POST /activate` in v6

* Add a changelog on the API change

* remove everything V5 related from wire-api

* updated brig after removing V5 types

* Adjust expectation in phone v5 test

---------

Co-authored-by: Leif Battermann <leif.battermann@wire.com>