RabbitMQ temporary clients (#4360)
Implements "temporary" queues for listening to events not bound to a specific client, meant to be used by team management or similar services.

When GET /events is called without a client_id parameter, we create a new temporary queue and bind it to the user-notifications exchange with routing keys <user-id> and <user-id>.temporary.

When a notification is published to RabbitMQ to all clients of a user, nothing changes, and <user-id> is used as its routing key. When it is published to a list of clients, it is now also published with routing key <user-id>.temporary. Each notifications is only published once with the <user-id>.temporary routing key even if the user has multiple capable clients.

When the websocket is closed, the temporary queue is deleted.


---------

Co-authored-by: Sven Tennie <sven.tennie@wire.com>