gundeck/cannon: Fixup for temporary recipients of events (#4379)
* gundeck: Allow temporary recipients to receive events even when the user doesn't have any regular RabbitMQ clients

* integration: Add test to ensure temp notif queues are deleted

Also:
- Extract function to create RabbitMQAdminClient for a backend resource
- Extract function to runDynamicBackends returning the BackendResources instead
of returning just the domains.

* cannon: Use exclusive auto-delete queues for temp clients

Also:

* Use Data.Unique.Unqiue to keep track of channels instead of (ByteString,
ByteString) because it is way less fuss to generate the Unqiue and it doesn't
really need to be (UserId, ClientId) tuple. This also removes polymorphism for
the RabbitMqPool type.

* Log conneciton close while draining at Debug level instead of Info

* integration: Make tests less flaky by expecting races

The race in these tests is between the RabbitMQ exchange doing the fanout for
previous events and the test creating a temporary queue.