Allow upgrading from ES6 to OpenSearch (#4444)
This commit supports both, ElasticSearch 6.8 and OpenSearch 1.3. The heavy-lifting is done by using a new version of our Bloodhound fork. Please note that this fork is not guaranteed to support all operations; only those we're currently using.

To be able to test it, we provide both index searches in our test setups (docker-compose and Kubernetes.)

---------

Co-authored-by: Sven Tennie <sven.tennie@wire.com>