Rate limit password hashing operations (#4353)
* Put all password hashing operations in one place

Once they are all in that effect, we can start asking for either the user ID or
the IP address and use that to rate limit.

Changes:
1. Move verifyPassword and verifyPasswordWithStatus to wire-subsystems
2. verifyPasswordWithStatus doesn't ask password to be updated if the current
config says that the password should be hashed with Scrypt.
3. Move PasswordStatus type to wire-subsystems as it is not part of the API.
4. Use Wire.Sem.Random (and hence openssl in prod) to generate salt.
5. Move all password tests from wire-api to wire-subsystems.
6. Use HashPassword effect for password verification everywhere.

* Implement effect to check for rate limits

* MockInterpreters: Update imports

MockInterpreters.ActivationCodeStore: Rename code -> emailKeyToCode to avoid
name collisions

* HashPassword: Export only rate limited helpers to invoke effect actions

* RateLimit: Allow exceptions

* charts/{brig,galley}: Add config options for rate limiting password hashing

* integration: Fix failing tests due to X-Forwarded-For and password hashing rate limits

* Docs and changelog

* brig-integration: Fix failing tests due to X-Forwarded-For and password hashing rate limits

* hack/helm_vars/wire-server: Don't rate limit tests

* brig-integration: Add X-Forwarded-For in calls that need it

* brig: Fix JSON convertion for RichError

The ToJSON instance does it by merging WaiError and the rich object's JSON and
ensuring that an Object is returned. Ignoring WaiError ends up as an empty array
because most of the RichErrors use `()`.

* Move testLimitRetries to integration

Use backend with config so password hashing rate limiting doesn't get in the way.

* integration: Add response headers to failure context

* brig-integration: Remove silly X-Forwarded-For headers

* Move testTooManyCookies to integration

Also:

Use getCurrentTime for recording cookie creation time instead of currentTime IO
from Brig's Env which updates every second. This makes writing tests much
easier. It is anyway not clear why we'd want the time to be updated like that.

* integration/setupProvider: Always use randomly generated email

* integration: Move test for searching for whitelisted services

* integration/setupProvider: Return result of `getProvider` instead of `newProvider`

This way all provider details are in the returned object, except for newly
created password. If a test needs this password, it should set it in the
`NewProvider` object.

* provider-password-reset: RateLimit by provider ID instead of IP Address

* Move provider password reset tests to new integration suite

* Move password change test to integration suite

Leaves the test for changing password for anon user in the old test suite,
because its too much work to integrate that into the new test suite.

* brig-integration: Remove duplicated line in test

This gets also gets rid of an error due to logging in too much causing rate
limits to kick in.

* stern-integration: Add X-Forawrded-For when regsitering users

* galley-integration: Add X-Forwarded-For in register and code-check

* Move test for checking scim token limit to new integration suite

* spar-integration: set X-Forwarded-For when registering users

* RateLimit: Add info log when doRateLimit fails

This also moves the doRateLimit inside the interpreter and hence gets rid of
`Error RateLimitExceeded` constraint from most places.

---------

Co-authored-by: Stefan Berthold <stefan.berthold@wire.com>