Remove TokenPair type class
The 'newAccessToken' function is moved to the 'AccessTokenLike' typeclass
returns an 'AccessToken' type which contains the serialized token, removing the
need for polymorphism.

Similarly 'renewAccessToken' function in the 'UserTokenLike' typeclass returns
an 'AccessToken'.

We might be able to get rid of all of this polymophism as things are not _so_
complicated. This is just the first step.