zauth: Unify User/LegalHoldUser and Acess/LegalHoldAccess using a phantom type paramter
These are essentially same things and we have invented type classes to treat
them so, but if they contain the same data structures inside and have same
serialisation/de-serialisation logic, there is little point in having new type
wrappers for LegalHoldUser and LegalHoldAccess.

The phantom type parameter is of type `UserTokenType` with values `ActualUser`
and `LHUser`. These names don't completely make any sense because there is no
such thing as a `LegalHoldUser` different from `ActualUser`. It _should_ rather
be called something like `TokenType`, but that name is taken by something else.
I suspect this type will go away soon, so while transitioning its probably fine
to keep it as it.

Doing all this makes it easier for The typeclasses `AccessTokenLike` and
`UserTokenLike` to be merged. The actual use of those type classes is to
determine which ttls to apply to these the `UserTokenType`, which can now be
made clearer in the code.