zauth: Use 'Type' as the parameter for the 'Token' type instead of body
Create a type family 'Body t' which defines the body based on the type of the
token. This removes illegal states like token of type Bot having a User as the
body.

As a result the token creation can be greatly simplified with it being passed
the parameters for creating the header and the body of correct type. This helps
reduce some duplication between ZAuth.Creation and
Wire.AuthenticationSubsystems.ZAuth. This also pushed the need for the random
number generation to the subsystems code making the creation even lighter. Maybe
we can get rid of polysemy from there too making the whole thing very simple.

The classes 'UserTokenLike' and 'AccessTokenLike' also become much simpler as
they can just have a general constraint like '(Body t ~ User)' and '(Body t ~
Access) respectively. Perhaps these classes can be gotten rid of once we can use
some state from the AuthenticationSubsystem.