[WPB-16420] Move saml2-web-sso repo to /libs/ (#4492)
* Import saml2-web-sso as library

The idea to host it as an independent library never has never really
taken off...

* saml2-web-sso: Formatting is done by surrounding treeformat

* saml2-web-sso: Delete old project setup / Reformat cabal file

* saml2-web-sso: Add to wire-server Cabal config

* saml2-web-sso: Format cabal file

* Fix: example SAML SP had an undesired hole.

* make sanitize-pr

* Tweak README

* rm config file from ancient forgotten dev setup.

* Run saml2-web-sso tests

Tests can be executed now, so we should.

* No constraint on the saml2-web-sso version

We simply want to take what's in this current checkout.

* Delete .gitignore

Let's start with what we already ignored at top-level. If we ever need
special rules here, we can simply create a new .gitignore file.

* Add changelog

* Reformat

* Fix Hlint issues that cannot be auto-fixed

* More prominent warnings when hlint-inplace fails.

* Run tests without SAML2_WEB_SSO_ROOT env var

Having to define this var just makes things complicated.

* Use camelCase to increase hlint's happiness

* Re-generate default.nix

---------

Co-authored-by: Sven Tennie <sven.tennie@wire.com>